+++ src/Store/Store.php
@@ -143,7 +143,7 @@ class Store implements StoreInterface
     ) {
         return $this
             ->adapterFor($record)
-            ->related($relationshipName)
+            ->getRelated($relationshipName)
             ->query($record, $params);
     }
 
@@ -157,7 +157,7 @@ class Store implements StoreInterface
     ) {
         return $this
             ->adapterFor($record)
-            ->related($relationshipName)
+            ->getRelated($relationshipName)
             ->relationship($record, $params);
     }
 
@@ -172,7 +172,7 @@ class Store implements StoreInterface
     ) {
         return $this
             ->adapterFor($record)
-            ->related($relationshipName)
+            ->getRelated($relationshipName)
             ->replace($record, $relationship, $params);
     }
 
@@ -308,7 +308,7 @@ class Store implements StoreInterface
      */
     private function adapterForHasMany($resourceType, $relationshipName)
     {
-        $adapter = $this->adapterFor($resourceType)->related($relationshipName);
+        $adapter = $this->adapterFor($resourceType)->getRelated($relationshipName);
 
         if (!$adapter instanceof HasManyAdapterInterface) {
             throw new RuntimeException("Expecting a has-many relationship adapter.");